+++ tests/test_blink_functions.py
@@ -2,8 +2,8 @@
 
 import unittest
 from unittest import mock
-import random
-from blinkpy import blinkpy as blinkpy
+from blinkpy import blinkpy
+from blinkpy.helpers.constants import BLINK_URL
 import tests.mock_responses as mresp
 
 USERNAME = 'foobar'
@@ -17,180 +17,107 @@ class TestBlinkFunctions(unittest.TestCase):
         """Set up Blink module."""
         self.blink = blinkpy.Blink(username=USERNAME,
                                    password=PASSWORD)
-        (self.region_id, self.region), = mresp.LOGIN_RESPONSE['region'].items()
-        self.test_urls = blinkpy.BlinkURLHandler(self.region_id)
-        self.urls = self.test_urls
+        # pylint: disable=protected-access
+        self.blink._auth_header = {
+            'Host': 'test.url.tld',
+            'TOKEN_AUTH': 'foobar123'
+        }
+        self.blink.urls = blinkpy.BlinkURLHandler('test')
+        self.config = {
+            'device_id': 1111,
+            'name': 'foobar',
+            'armed': False,
+            'thumbnail': '/test',
+            'video': '/test.mp4',
+            'temp': 80,
+            'battery': 3,
+            'notifications': 2,
+            'region_id': 'test',
+            'device_type': 'camera'
+        }
+        self.camera = blinkpy.BlinkCamera(self.config, self.blink)
 
     def tearDown(self):
         """Clean up after test."""
         self.blink = None
-        self.region = None
-        self.region_id = None
-        self.test_urls = None
-        self.urls = None
+        self.config = {}
+        self.camera = None
 
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_set_motion_detect(self, mock_get, mock_post):
-        """Checks if we can set motion detection."""
-        self.blink.setup_system()
-        self.test_urls = blinkpy.BlinkURLHandler(self.region_id)
-        test_cameras = mresp.get_test_cameras(self.test_urls.base_url)
-        for camera_name in test_cameras:
-            self.blink.cameras[camera_name].set_motion_detect(True)
-            self.blink.refresh()
-            self.assertEqual(self.blink.cameras[camera_name].armed, True)
-            self.blink.cameras[camera_name].set_motion_detect(False)
-            self.blink.refresh()
-            self.assertEqual(self.blink.cameras[camera_name].armed, False)
+    @mock.patch('blinkpy.blinkpy._request')
+    def test_get_videos(self, req):
+        """Test video access."""
+        req.return_value = [
+            {
+                'camera_name': 'foobar',
+                'address': '/new/test.mp4',
+                'thumbnail': '/test/thumb'
+            }
+        ]
+        self.blink.get_videos()
+        self.assertEqual(self.blink.videos['foobar'][0]['clip'],
+                         '/new/test.mp4')
+        self.assertEqual(self.blink.videos['foobar'][0]['thumb'],
+                         '/test/thumb')
 
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_last_motion(self, mock_get, mock_post):
-        """Checks that we can get the last motion info."""
-        self.test_urls = blinkpy.BlinkURLHandler(self.region_id)
-        test_events = mresp.RESPONSE['event']
-        test_video = dict()
-        test_image = dict()
-        test_time = dict()
-        for event in test_events:
-            if event['type'] == 'motion':
-                url = self.test_urls.base_url + event['video_url']
-                test_video[event['camera_name']] = url
-                test_image[event['camera_name']] = url[:-3] + 'jpg'
-                test_time[event['camera_name']] = event['created_at']
+    @mock.patch('blinkpy.blinkpy._request')
+    def test_get_cameras(self, req):
+        """Test camera extraction."""
+        req.return_value = {'devices': [self.config]}
+        self.blink.get_cameras()
+        self.assertTrue('foobar' in self.blink.cameras)
 
-        self.blink.setup_system()
-        for name in self.blink.cameras:
-            camera = self.blink.cameras[name]
-            self.blink.last_motion()
-            if name in test_video:
-                self.assertEqual(camera.motion['video'], test_video[name])
-            else:
-                self.assertEqual(camera.motion, {})
-            if name in test_image:
-                self.assertEqual(camera.motion['image'], test_image[name])
-            else:
-                self.assertEqual(camera.motion, {})
-            if name in test_video:
-                self.assertEqual(camera.motion['time'], test_time[name])
-            else:
-                self.assertEqual(camera.motion, {})
+    @mock.patch('blinkpy.blinkpy._request')
+    def test_image_refresh(self, req):
+        """Test image refresh function."""
+        req.return_value = {'devices': [self.config]}
+        image = self.camera.image_refresh()
+        self.assertEqual(image,
+                         'https://rest.test.{}/test.jpg'.format(BLINK_URL))
 
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_take_new_picture(self, mock_get, mock_post):
-        """Checks if we can take a new picture and retrieve the thumbnail."""
-        self.blink.setup_system()
-        test_cameras = mresp.get_test_cameras(self.test_urls.base_url)
-        test_thumbnail = self.test_urls.base_url + mresp.NEW_THUMBNAIL + '.jpg'
-        # Snap picture for each camera and check new thumb
-        for camera_name in test_cameras:
-            camera = self.blink.cameras[camera_name]
-            camera.snap_picture()
-            camera.image_refresh()
-            self.assertEqual(camera.thumbnail, test_thumbnail)
+    @mock.patch('blinkpy.blinkpy._request')
+    def test_video_count(self, req):
+        """Test video count function."""
+        req.return_value = {'count': 1}
+        self.assertEqual(self.blink.video_count, 1)
 
-        # Manually set thumbnail, and then globally refresh and check
-        for camera_name in test_cameras:
-            camera = self.blink.cameras[camera_name]
-            camera.thumbnail = 'Testing'
-            self.assertEqual(camera.thumbnail, 'Testing')
-
+    @mock.patch('blinkpy.blinkpy._request')
+    @mock.patch('blinkpy.blinkpy.Blink._video_request')
+    def test_refresh(self, vid_req, req):
+        """Test blinkpy refresh function."""
+        self.blink.cameras = {'foobar': self.camera}
+        req.return_value = {'devices': [self.config]}
+        vid_req.return_value = [
+            {
+                'camera_name': 'foobar',
+                'address': '/new.mp4',
+                'thumbnail': '/new'
+            }
+        ]
         self.blink.refresh()
-        for camera_name in test_cameras:
-            camera = self.blink.cameras[camera_name]
-            self.assertEqual(camera.thumbnail, test_thumbnail)
+        test_camera = self.blink.cameras['foobar']
+        self.assertEqual(test_camera.clip,
+                         'https://rest.test.{}/new.mp4'.format(BLINK_URL))
+        self.assertEqual(test_camera.thumbnail,
+                         'https://rest.test.{}/new.jpg'.format(BLINK_URL))
 
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_image_with_bad_data(self, mock_get, mock_post):
-        """Checks for handling of bad keys."""
-        self.blink.setup_system()
-        for camera_name in self.blink.cameras:
-            camera = self.blink.cameras[camera_name]
-            camera.snap_picture()
-            camera.urls.home_url = "use_bad_response"
-            self.assertEqual(camera.image_refresh(), None)
+    def test_set_links(self):
+        """Test the link set method."""
+        self.blink.cameras = {'foobar': self.camera}
+        self.blink.network_id = 9999
+        self.blink.set_links()
+        net_url = "{}/{}".format(self.blink.urls.network_url, 9999)
+        self.assertEqual(self.camera.image_link,
+                         "{}/camera/1111/thumbnail".format(net_url))
+        self.assertEqual(self.camera.arm_link,
+                         "{}/camera/1111/".format(net_url))
 
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_camera_random_case(self, mock_get, mock_post):
-        """Checks for case of camera name."""
-        self.blink.setup_system()
-        for camera_name in self.blink.cameras:
-
-            rand_name = camera_name
-            # Make sure we never pass this test if rand_name = camera_name
-            while rand_name == camera_name:
-                rand_name = ''.join(
-                    random.choice(
-                        (str.upper, str.lower)
-                    )(x) for x in camera_name)
-
-            self.assertEqual(self.blink.cameras[camera_name].name,
-                             self.blink.cameras[rand_name].name)
-
-    def test_camera_update(self):
-        """Checks that the update function is doing the right thing."""
-        self.test_urls = blinkpy.BlinkURLHandler('test')
-        test_config = mresp.FIRST_CAMERA
-        self.urls = self.test_urls
-        test_camera = blinkpy.BlinkCamera(test_config, self)
-        test_update = mresp.SECOND_CAMERA
-        test_camera.update(test_update)
-        test_image_url = self.test_urls.base_url + test_update['thumbnail']
-        test_thumbnail = test_image_url + '.jpg'
-        test_clip = test_image_url + '.mp4'
-        self.assertEqual(test_camera.name, test_update['name'])
-        self.assertEqual(test_camera.armed, test_update['armed'])
-        self.assertEqual(test_camera.thumbnail, test_thumbnail)
-        self.assertEqual(test_camera.clip, test_clip)
-        self.assertEqual(test_camera.temperature, test_update['temp'])
-        self.assertEqual(test_camera.battery, test_update['battery'])
-        self.assertEqual(test_camera.notifications,
-                         test_update['notifications'])
-
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_camera_thumbs(self, mock_get, mock_post):
-        """Checks to see if we can retrieve camera thumbs."""
-        self.test_urls = blinkpy.BlinkURLHandler(self.region_id)
-        test_cameras = mresp.get_test_cameras(self.test_urls.base_url)
-        self.blink.setup_system()
-        for name in self.blink.cameras:
-            thumb = self.blink.camera_thumbs[name]
-            self.assertEqual(test_cameras[name]['thumbnail'], thumb)
-
-    @mock.patch('blinkpy.blinkpy.copyfileobj',
-                side_effect=mresp.mocked_copyfileobj)
-    @mock.patch('blinkpy.blinkpy.requests.post',
-                side_effect=mresp.mocked_requests_post)
-    @mock.patch('blinkpy.blinkpy.requests.get',
-                side_effect=mresp.mocked_requests_get)
-    def test_image_to_file(self, mock_get, mock_post, mock_copyfileobj):
-        """Checks that we can write an image to file."""
-        self.blink.setup_system()
-        cameras = self.blink.cameras
-        filename = '/tmp/test.jpg'
-        test_files = list()
-        for camera_name in cameras:
-            camera = cameras[camera_name]
-            test_files.append(mresp.MOCK_BYTES)
-            mock_fh = mock.mock_open()
-            with mock.patch('builtins.open', mock_fh, create=True):
-                camera.image_to_file(camera_name + filename)
-            mock_fh.assert_called_once_with(camera_name + filename, 'wb')
-        self.assertEqual(test_files, mresp.FAKE_FILES)
+    @mock.patch('blinkpy.blinkpy._request')
+    def test_backup_url(self, req):
+        """Test backup login method."""
+        req.side_effect = [
+            mresp.mocked_requests_post(None),
+            {'authtoken': {'authtoken': 'foobar123'}}
+        ]
+        self.blink.get_auth_token()
+        self.assertEqual(self.blink.region_id, 'piri')
+        self.assertEqual(self.blink.region, 'UNKNOWN')